+++ src/com/kaylerrenslow/armaDialogCreator/gui/fx/main/actions/mainMenu/file/FileSaveAction.java
@@ -10,11 +10,14 @@
 
 package com.kaylerrenslow.armaDialogCreator.gui.fx.main.actions.mainMenu.file;
 
-import com.kaylerrenslow.armaDialogCreator.main.ArmaDialogCreator;
-import com.kaylerrenslow.armaDialogCreator.main.ExceptionHandler;
+import com.kaylerrenslow.armaDialogCreator.data.ApplicationDataManager;
+import com.kaylerrenslow.armaDialogCreator.gui.fx.notification.Notification;
+import com.kaylerrenslow.armaDialogCreator.gui.fx.notification.Notifications;
+import com.kaylerrenslow.armaDialogCreator.main.Lang;
 import javafx.event.ActionEvent;
 import javafx.event.EventHandler;
 
+import java.io.File;
 import java.io.IOException;
 
 /**
@@ -23,10 +26,24 @@ import java.io.IOException;
 public class FileSaveAction implements EventHandler<ActionEvent> {
 	@Override
 	public void handle(ActionEvent event) {
+		Notification resultNotification;
+
 		try {
-			ArmaDialogCreator.getApplicationDataManager().saveProject();
+			ApplicationDataManager.getInstance().saveProject();
+			File saveFile = ApplicationDataManager.getInstance().getCurrentProject().getProjectSaveDirectory();
+			resultNotification = new Notification(
+					Lang.ApplicationBundle().getString("Notifications.ProjectSave.Success.notification_title"),
+					String.format(Lang.ApplicationBundle().getString("Notifications.ProjectSave.Success.notification_body_f"), saveFile.getAbsolutePath())
+			);
 		} catch (IOException e) {
-			ExceptionHandler.error(e);
+			e.printStackTrace();
+			String reason = e.getMessage() != null && e.getMessage().length() > 0 ? e.getMessage() : Lang.ApplicationBundle().getString("Notifications.ProjectSave.Fail.unknown_reason");
+			resultNotification = new Notification(
+					Lang.ApplicationBundle().getString("Notifications.ProjectSave.Fail.notification_title"),
+					String.format(Lang.ApplicationBundle().getString("Notifications.ProjectSave.Fail.notification_body_f"), reason),
+					10 * 1000, true
+			);
 		}
+		Notifications.showNotification(resultNotification);
 	}
 }