+++ src/com/kaylerrenslow/armaDialogCreator/gui/fx/notification/Notifications.java
@@ -11,6 +11,7 @@
 package com.kaylerrenslow.armaDialogCreator.gui.fx.notification;
 
 import com.kaylerrenslow.armaDialogCreator.main.ArmaDialogCreator;
+import com.kaylerrenslow.armaDialogCreator.util.ReadOnlyList;
 import javafx.application.Platform;
 import javafx.concurrent.Task;
 import org.jetbrains.annotations.NotNull;
@@ -18,42 +19,75 @@ import org.jetbrains.annotations.NotNull;
 import java.util.LinkedList;
 
 /**
+ Handles all {@link Notification} instances that are to be displayed. When a {@link Notification} is shown via {@link #showNotification(Notification)}, {@link #getPastNotifications()} will have
+ that notification appended and the notification will be displayed.
+
  @author Kayler
  @since 11/16/2016 */
 public class Notifications {
-	private static LinkedList<NotificationWrapper> showingNotifications = new LinkedList<>();
-	private static NotificationsVisibilityTask visibilityTask = new NotificationsVisibilityTask();
+	private static final Notifications INSTANCE = new Notifications();
+	private static final int MAX_HISTORY_NOTIFICATIONS = 15;
 
+
+	/**
+	 Shows the specified {@link Notification}
+
+	 @param notification notification to show
+	 */
 	public static void showNotification(@NotNull Notification notification) {
-		showingNotifications.add(new NotificationWrapper(notification, System.currentTimeMillis()));
-		ArmaDialogCreator.getMainWindow().getNotificationPane().addNotification(notification);
-		if (!visibilityTask.isRunning()) {
-			new Thread(visibilityTask).start();
-		}
+		INSTANCE.doShowNotification(notification);
 	}
 
-	private static class NotificationWrapper {
-		private final Notification notification;
-		private final long timeShown;
+	/** Return a list of past notifications that were displayed. */
+	@NotNull
+	public static ReadOnlyList<NotificationDescriptor> getPastNotifications() {
+		return INSTANCE.pastNotificationsReadOnly;
+	}
 
-		public NotificationWrapper(@NotNull Notification notification, long timeShown) {
-			this.notification = notification;
-			this.timeShown = timeShown;
+
+	private Notifications() {
+	}
+
+
+	private final LinkedList<NotificationDescriptor> showingNotifications = new LinkedList<>();
+	private final NotificationsVisibilityTask visibilityTask = new NotificationsVisibilityTask(this);
+	private final LinkedList<NotificationDescriptor> pastNotifications = new LinkedList<>();
+	private final ReadOnlyList<NotificationDescriptor> pastNotificationsReadOnly = new ReadOnlyList<>(pastNotifications);
+
+	private void doShowNotification(@NotNull Notification notification) {
+		NotificationDescriptor descriptor = new NotificationDescriptor(notification, System.currentTimeMillis());
+		pastNotifications.add(descriptor);
+		if (pastNotifications.size() >= MAX_HISTORY_NOTIFICATIONS) {
+			while (pastNotifications.size() >= MAX_HISTORY_NOTIFICATIONS) {
+				pastNotifications.removeFirst();
+			}
+		}
+		showingNotifications.add(descriptor);
+		ArmaDialogCreator.getMainWindow().getNotificationPane().addNotification(notification);
+		if (!visibilityTask.isRunning()) {
+			Thread thread = new Thread(visibilityTask);
+			thread.setDaemon(true);
+			thread.start();
 		}
 	}
 
 	private static class NotificationsVisibilityTask extends Task<Boolean> {
 
+		private final Notifications notifications;
+
+		public NotificationsVisibilityTask(Notifications notifications) {
+			this.notifications = notifications;
+		}
 
 		@Override
 		protected Boolean call() throws Exception {
-			LinkedList<NotificationWrapper> tohide = new LinkedList<>();
+			LinkedList<NotificationDescriptor> tohide = new LinkedList<>();
 			while (true) {
 				long now = System.currentTimeMillis();
-				for (int i = 0; i < showingNotifications.size(); ) {
-					NotificationWrapper wrapper = showingNotifications.get(i);
-					if (!wrapper.notification.isShowing() || wrapper.timeShown + wrapper.notification.getDisplayDurationMilliseconds() <= now) {
-						NotificationWrapper remove = showingNotifications.remove(i);
+				for (int i = 0; i < notifications.showingNotifications.size(); ) {
+					NotificationDescriptor wrapper = notifications.showingNotifications.get(i);
+					if (!wrapper.getNotification().isShowing() || wrapper.getTimeShown() + wrapper.getNotification().getDisplayDurationMilliseconds() <= now) {
+						NotificationDescriptor remove = notifications.showingNotifications.remove(i);
 						tohide.add(remove);
 						continue;
 					}
@@ -64,20 +98,20 @@ public class Notifications {
 						@Override
 						public void run() {
 							while (tohide.size() > 0) {
-								tohide.removeFirst().notification.setShowing(false);
+								tohide.removeFirst().getNotification().setShowing(false);
 							}
 						}
 					});
 				}
-				if (showingNotifications.size() == 0) {
+				if (notifications.showingNotifications.size() == 0) {
 					try {
 						Thread.sleep(2000);
-					} catch (Exception e) {
+					} catch (Exception ignore) {
 					}
 				} else {
 					try {
 						Thread.sleep(20);
-					} catch (Exception e) {
+					} catch (Exception ignore) {
 					}
 				}
 			}